Revert "Revert "update convertlead""
This reverts commit 834ddbf37a9568253022255cf955ee5199378d37.

+++ backend/app/Http/Controllers/Api/Management/Agency/CampaignController.php
@@ -106,9 +106,13 @@ class CampaignController extends Controller
             ->user()
             ->getCompanyBy($company)
             ->getDealBy($deal)
-            ->getCampaignBy($id);
-
-        $campaign->delete();
+            ->campaigns()->withTrashed()->where('id', $id)->first();
+             
+        if($campaign->trashed()) {
+            $campaign->restore();
+        } else {
+            $campaign->delete();
+        }
 
         return $campaign;
     }
